Created new Installer module and moved everything from CLI there
