+++ composer.json
@@ -61,7 +61,8 @@
             "Shlinkio\\Shlink\\CLI\\": "module/CLI/src",
             "Shlinkio\\Shlink\\Rest\\": "module/Rest/src",
             "Shlinkio\\Shlink\\Core\\": "module/Core/src",
-            "Shlinkio\\Shlink\\Common\\": "module/Common/src"
+            "Shlinkio\\Shlink\\Common\\": "module/Common/src",
+            "Shlinkio\\Shlink\\Installer\\": "module/Installer/src"
         },
         "files": [
             "module/Common/functions/functions.php"
@@ -78,7 +79,8 @@
             "ShlinkioTest\\Shlink\\Common\\": [
                 "module/Common/test",
                 "module/Common/test-func"
-            ]
+            ],
+            "ShlinkioTest\\Shlink\\Installer\\": "module/Installer/test"
         }
     },
     "scripts": {
@@ -101,8 +103,8 @@
             "phpcov merge build --html build/html"
         ],
         "stan": "phpstan analyse module/*/src/ --level=6 -c phpstan.neon",
-        "infect": "infection --threads=4 --min-msi=55 --only-covered --log-verbosity=2",
-        "infect-show": "infection --threads=4 --min-msi=55 --only-covered --log-verbosity=2 --show-mutations",
+        "infect": "infection --threads=4 --min-msi=60 --only-covered --log-verbosity=2",
+        "infect-show": "infection --threads=4 --min-msi=60 --only-covered --log-verbosity=2 --show-mutations",
         "expressive": "expressive"
     },
     "config": {
