+++ module/Installer/config/translator.config.php
@@ -1,0 +1,16 @@
+<?php
+declare(strict_types=1);
+
+return [
+
+    'translator' => [
+        'translation_file_patterns' => [
+            [
+                'type' => 'gettext',
+                'base_dir' => __DIR__ . '/../lang',
+                'pattern' => '%s.mo',
+            ],
+        ],
+    ],
+
+];
