+++ module/Installer/src/Config/ConfigCustomizerManager.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Shlinkio\Shlink\CLI\Install;
+namespace Shlinkio\Shlink\Installer\Config;
 
-use Shlinkio\Shlink\CLI\Install\Plugin\ConfigCustomizerInterface;
+use Shlinkio\Shlink\Installer\Config\Plugin\ConfigCustomizerInterface;
 use Zend\ServiceManager\AbstractPluginManager;
 
 class ConfigCustomizerManager extends AbstractPluginManager implements ConfigCustomizerManagerInterface
