+++ module/Installer/test/Config/Plugin/LanguageConfigCustomizerTest.php
@@ -1,32 +1,31 @@
 <?php
 declare(strict_types=1);
 
-namespace ShlinkioTest\Shlink\CLI\Install\Plugin;
+namespace ShlinkioTest\Shlink\Installer\Config\Plugin;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
-use Shlinkio\Shlink\CLI\Install\Plugin\ApplicationConfigCustomizer;
-use Shlinkio\Shlink\CLI\Model\CustomizableAppConfig;
+use Shlinkio\Shlink\Installer\Config\Plugin\LanguageConfigCustomizer;
+use Shlinkio\Shlink\Installer\Model\CustomizableAppConfig;
 use Symfony\Component\Console\Style\SymfonyStyle;
 
-class ApplicationConfigCustomizerTest extends TestCase
+class LanguageConfigCustomizerTest extends TestCase
 {
     /**
-     * @var ApplicationConfigCustomizer
+     * @var LanguageConfigCustomizer
      */
-    private $plugin;
+    protected $plugin;
     /**
      * @var ObjectProphecy
      */
-    private $io;
+    protected $io;
 
     public function setUp()
     {
         $this->io = $this->prophesize(SymfonyStyle::class);
         $this->io->title(Argument::any())->willReturn(null);
-
-        $this->plugin = new ApplicationConfigCustomizer();
+        $this->plugin = new LanguageConfigCustomizer();
     }
 
     /**
@@ -34,16 +33,16 @@ class ApplicationConfigCustomizerTest extends TestCase
      */
     public function configIsRequestedToTheUser()
     {
-        $ask = $this->io->ask(Argument::cetera())->willReturn('the_secret');
+        $ask = $this->io->choice(Argument::cetera())->willReturn('en');
         $config = new CustomizableAppConfig();
 
         $this->plugin->process($this->io->reveal(), $config);
 
-        $this->assertTrue($config->hasApp());
+        $this->assertTrue($config->hasLanguage());
         $this->assertEquals([
-            'SECRET' => 'the_secret',
-            'DISABLE_TRACK_PARAM' => 'the_secret',
-        ], $config->getApp());
+            'DEFAULT' => 'en',
+            'CLI' => 'en',
+        ], $config->getLanguage());
         $ask->shouldHaveBeenCalledTimes(2);
     }
 
@@ -52,20 +51,21 @@ class ApplicationConfigCustomizerTest extends TestCase
      */
     public function overwriteIsRequestedIfValueIsAlreadySet()
     {
-        $ask = $this->io->ask(Argument::cetera())->willReturn('the_new_secret');
+        $choice = $this->io->choice(Argument::cetera())->willReturn('es');
         $confirm = $this->io->confirm(Argument::cetera())->willReturn(false);
         $config = new CustomizableAppConfig();
-        $config->setApp([
-            'SECRET' => 'foo',
+        $config->setLanguage([
+            'DEFAULT' => 'en',
+            'CLI' => 'en',
         ]);
 
         $this->plugin->process($this->io->reveal(), $config);
 
         $this->assertEquals([
-            'SECRET' => 'the_new_secret',
-            'DISABLE_TRACK_PARAM' => 'the_new_secret',
-        ], $config->getApp());
-        $ask->shouldHaveBeenCalledTimes(2);
+            'DEFAULT' => 'es',
+            'CLI' => 'es',
+        ], $config->getLanguage());
+        $choice->shouldHaveBeenCalledTimes(2);
         $confirm->shouldHaveBeenCalledTimes(1);
     }
 
@@ -74,18 +74,20 @@ class ApplicationConfigCustomizerTest extends TestCase
      */
     public function existingValueIsKeptIfRequested()
     {
-        $confirm = $this->io->confirm(Argument::cetera())->willReturn(true);
+        $ask = $this->io->confirm(Argument::cetera())->willReturn(true);
 
         $config = new CustomizableAppConfig();
-        $config->setApp([
-            'SECRET' => 'foo',
+        $config->setLanguage([
+            'DEFAULT' => 'es',
+            'CLI' => 'es',
         ]);
 
         $this->plugin->process($this->io->reveal(), $config);
 
         $this->assertEquals([
-            'SECRET' => 'foo',
-        ], $config->getApp());
-        $confirm->shouldHaveBeenCalledTimes(1);
+            'DEFAULT' => 'es',
+            'CLI' => 'es',
+        ], $config->getLanguage());
+        $ask->shouldHaveBeenCalledTimes(1);
     }
 }
